More fixes of comparing int to size_t

+++ cpp/src/geo.cpp
@@ -938,7 +938,7 @@ void kerr_geo_kepler_parameters(double &p, double &e, double &x, double a, doubl
 }
 
 void kerr_geo_kepler_parameters(int n, double* p, double* e, double* x, const double* a, const double* En, const double* Lz, const double* Qc){
-	for(size_t i = 0; i < n; i++){
+	for(int i = 0; i < n; i++){
 		kerr_geo_kepler_parameters(p[i], e[i], x[i], a[i], En[i], Lz[i], Qc[i]);
 	}
 }
@@ -998,7 +998,7 @@ void kerr_geo_orbital_constants(double &En, double &Lz, double &Qc, double a, do
 }
 
 void kerr_geo_orbital_constants(int n, double* En, double* Lz, double* Qc, const double* a, const double* p, const double* e, const double* x){
-	for(size_t i = 0; i < n; i++){
+	for(int i = 0; i < n; i++){
 		kerr_geo_orbital_constants(En[i], Lz[i], Qc[i], a[i], p[i], e[i], x[i]);
 	}
 }
